+++ tests/test_cases.cc
@@ -255,6 +255,7 @@ using namespace clanguml::test::matchers;
 #include "t00053/test_case.h"
 #include "t00054/test_case.h"
 #include "t00055/test_case.h"
+#include "t00057/test_case.h"
 
 ///
 /// Sequence diagram tests