+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+  * Added support for plain C11 translation units (#97)
   * Added 'row' and 'column' layout hints for aligning elements (#90)
   * Added 'together' layout hint for grouping elements (#43)
   * Enabled adding notes to class methods and members (#87)